GptResponseの必須フィールドを元に戻し、エラー判定を分離
成功時のレスポンス構造まで過剰にnullable化していたため、GptResponseは
元の必須フィールド構成に戻した。OpenAIのエラーレスポンス判定は専用の
GptErrorResponse型とJSONの事前チェックで行うようにした。
