+++ src/commonMain/kotlin/net/matsudamper/gptclient/client/openai/GptResponse.kt
@@ -6,24 +6,15 @@ import net.matsudamper.gptclient.serialization.StringEnum
 
 @Serializable
 data class GptResponse(
-    @SerialName("id") val id: String? = null,
-    @SerialName("object") val objectType: String? = null,
-    @SerialName("created") val created: Long? = null,
-    @SerialName("model") val model: String? = null,
-    @SerialName("choices") val choices: List<Choice> = emptyList(),
-    @SerialName("usage") val usage: Usage? = null,
-    @SerialName("service_tier") val serviceTier: String? = null,
-    @SerialName("system_fingerprint") val systemFingerprint: String? = null,
-    @SerialName("error") val error: Error? = null,
+    @SerialName("id") val id: String,
+    @SerialName("object") val objectType: String,
+    @SerialName("created") val created: Long,
+    @SerialName("model") val model: String,
+    @SerialName("choices") val choices: List<Choice>,
+    @SerialName("usage") val usage: Usage,
+    @SerialName("service_tier") val serviceTier: String,
+    @SerialName("system_fingerprint") val systemFingerprint: String?,
 ) {
-    @Serializable
-    data class Error(
-        @SerialName("message") val message: String? = null,
-        @SerialName("type") val type: String? = null,
-        @SerialName("param") val param: String? = null,
-        @SerialName("code") val code: String? = null,
-    )
-
     @Serializable
     data class Choice(
         @SerialName("index") val index: Int,
